Restore the GX texture-size signature and native copy arguments
Remove the C++-only u32 dimensions from GXGetTexBufferSize. Both C and
C++ now use the u16 dimensions defined by the SDK implementation in
src/gx/GXTexture.c. Remove intermediate int casts from the miasma
renderer so its float dimensions convert directly through that API.

Use the GX APIs own argument conversions for framebuffer-copy dimensions
and coordinates, and express destination offset/alignment in two steps.
This replaces redundant masks and casts with normal SDK calls.

Objdiff: GetBackBufferRect 95.47486% -> 97.765366%;
GetBackBufferRect2 91.43519% -> 99.90741%; graphic 98.92415% ->
99.27192%. All other function scores, including pppRenderMiasma, and
all data scores remain unchanged after rebuilding header dependents.

Validated with full source ninja build, graphic/pppMiasma objdiff,
relocated float/double payload checks, global report/data comparison,
and PAL checksum.

+++ src/pppMiasma.cpp
@@ -295,9 +295,9 @@ void pppRenderMiasma(pppMiasma* pppMiasma, pppMiasmaRenderStep* param_2, _pppCtr
     packedWork.bytes[2] = (u8)(work->m_position[2] >> 7);
     packedWork.bytes[3] = (u8)(work->m_position[3] >> 7);
 
-    i4TexSize = GXGetTexBufferSize((int)kPppAlignmentScaleScreenWidth, (int)kPppAlignmentScaleScreenHalfHeight,
+    i4TexSize = GXGetTexBufferSize(kPppAlignmentScaleScreenWidth, kPppAlignmentScaleScreenHalfHeight,
                                    (GXTexFmt)6, GX_FALSE, 0);
-    rgba8TexSize = GXGetTexBufferSize((int)kPppAlignmentScaleScreenWidth, (int)kPppAlignmentScaleScreenHalfHeight,
+    rgba8TexSize = GXGetTexBufferSize(kPppAlignmentScaleScreenWidth, kPppAlignmentScaleScreenHalfHeight,
                                       (GXTexFmt)0x28, GX_FALSE, 0);
 
     managerPos.x = ppvMng->m_matrix.value[0][3];
